Struct rc_box::ArcBox

#[repr(transparent)]pub struct ArcBox<T: ?Sized> { /* fields omitted */ }

Known unique version of [Arc].

This type is guaranteed to have the same repr as Box<T>. (The heap layout is that of Arc<T>.)

Implementations

impl ArcBox<dyn Any + 'static>

pub fn downcast<T>(self) -> Result<ArcBox<T>, Self> where
T: Any,

Attempt to downcast the box to a concrete type.

Examples

use std::any::Any;

fn print_if_string(value: ArcBox<dyn Any>) {
    if let Ok(string) = value.downcast::<String>() {
        println!("String ({}): {}", string.len(), string);
    }
}

let my_string = "Hello World".to_string();
let my_string: Arc<dyn Any> = Arc::new(my_string); 
print_if_string(my_string.try_into().unwrap());
let my_number: Arc<dyn Any> = Arc::new(0i8);
print_if_string(my_number.try_into().unwrap());

The unsizing as Arc is required until DST coercions are stabilized.

impl ArcBox<dyn Any + Send + 'static>

pub fn downcast<T>(self) -> Result<ArcBox<T>, Self> where
T: Any + Send,

Attempt to downcast the box to a concrete type.

Examples

use std::any::Any;

fn print_if_string(value: ArcBox<dyn Any + Send>) {
    if let Ok(string) = value.downcast::<String>() {
        println!("String ({}): {}", string.len(), string);
    }
}

let my_string = "Hello World".to_string();
let my_string: Arc<dyn Any + Send> = Arc::new(my_string); 
print_if_string(my_string.try_into().unwrap());
let my_number: Arc<dyn Any + Send> = Arc::new(0i8);
print_if_string(my_number.try_into().unwrap());

The unsizing as Arc is required until DST coercions are stabilized.

impl ArcBox<dyn Any + Send + Sync + 'static>

pub fn downcast<T>(self) -> Result<ArcBox<T>, Self> where
T: Any + Send + Sync,

Attempt to downcast the box to a concrete type.

Examples

use std::any::Any;

fn print_if_string(value: ArcBox<dyn Any + Send + Sync>) {
    if let Ok(string) = value.downcast::<String>() {
        println!("String ({}): {}", string.len(), string);
    }
}

let my_string = "Hello World".to_string();
let my_string: Arc<dyn Any + Send + Sync> = Arc::new(my_string); 
print_if_string(my_string.try_into().unwrap());
let my_number: Arc<dyn Any + Send + Sync> = Arc::new(0i8);
print_if_string(my_number.try_into().unwrap());

The unsizing as Arc is required until DST coercions are stabilized.

impl<T: ?Sized> ArcBox<T>

pub unsafe fn from_raw(ptr: *const T) -> Self

Construct an ArcBox from a raw pointer.

Safety

The raw pointer must have previously been acquired by a call to [ArcBox::into_raw], or [Arc::into_raw] where the Arc is known unique.

pub fn get_mut(this: &mut Self) -> Option<&mut T>

👎 Deprecated:
Use DerefMut instead

Get a mutable reference into the ArcBox.

This method exists only for API compatibility with Arc. Use DerefMut instead.

pub fn get_mut_unchecked(this: &mut Self) -> &mut T

👎 Deprecated:
Use DerefMut instead

Get a mutable reference into the ArcBox.

This method exists only for API compatibility with Arc. Use DerefMut instead.

pub fn as_raw(this: &Self) -> NonNull<T>

Returns a raw pointer to the object T pointed to by this ArcBox.

Note that this returns a [ptr::NonNull], not a raw pointer. That makes this function equivalent to as_raw_non_null.

pub fn into_raw(this: Self) -> NonNull<T>

Consume the ArcBox, returning the wrapped pointer.

To avoid a memory leak, the pointer must be converted back to a ArcBox, using [ArcBox::from_raw], or directly into a Arc, using [Arc::from_raw].

Note that this returns a [ptr::NonNull], not a raw pointer. That makes this function equivalent to into_raw_non_null.

pub fn leak<'a>(this: Self) -> &'a mut T where
T: 'a,

Consume and leak the ArcBox.

pub fn new(data: T) -> Self where
T: Sized,

Create a new ArcBox.

pub fn pin(x: T) -> Pin<ArcBox<T>> where
T: Sized,

Construct a new Pin<ArcBox<T>>. If T does not implement [Unpin], then the data will be pinned in memory and unable to be moved.

pub fn into_inner(this: Self) -> T where
T: Sized,

Deconstruct this ArcBox, returning the inner value.