rc_box
Struct RcBox

pub struct RcBox<T: ?Sized> { /* private fields */ }

Known unique version of Rc.

This type is guaranteed to have the same repr as Box<T>. (The heap layout is that of Rc<T>.)

Implementations

impl RcBox<dyn Any + 'static>

pub fn downcast<T>(self) -> Result<RcBox<T>, Self>
where T: Any,

Attempt to downcast the box to a concrete type.

Examples

use std::any::Any;

fn print_if_string(value: RcBox<dyn Any>) {
    if let Ok(string) = value.downcast::<String>() {
        println!("String ({}): {}", string.len(), string);
    }
}

let my_string = "Hello World".to_string();
let my_string: Rc<dyn Any> = Rc::new(my_string); 
print_if_string(my_string.try_into().unwrap());
let my_number: Rc<dyn Any> = Rc::new(0i8);
print_if_string(my_number.try_into().unwrap());

The unsizing as Rc is required until DST coercions are stabilized. Alternatively, activate the unsize feature to convert the pointer via an explicit method call.

impl RcBox<dyn Any + Send + 'static>

pub fn downcast<T>(self) -> Result<RcBox<T>, Self>
where T: Any + Send,

Attempt to downcast the box to a concrete type.

Examples

use std::any::Any;

fn print_if_string(value: RcBox<dyn Any + Send>) {
    if let Ok(string) = value.downcast::<String>() {
        println!("String ({}): {}", string.len(), string);
    }
}

let my_string = "Hello World".to_string();
let my_string: Rc<dyn Any + Send> = Rc::new(my_string); 
print_if_string(my_string.try_into().unwrap());
let my_number: Rc<dyn Any + Send> = Rc::new(0i8);
print_if_string(my_number.try_into().unwrap());

The unsizing as Rc is required until DST coercions are stabilized. Alternatively, activate the unsize feature to convert the pointer via an explicit method call.

impl RcBox<dyn Any + Send + Sync + 'static>

pub fn downcast<T>(self) -> Result<RcBox<T>, Self>
where T: Any + Send + Sync,

Attempt to downcast the box to a concrete type.

Examples

use std::any::Any;

fn print_if_string(value: RcBox<dyn Any + Send + Sync>) {
    if let Ok(string) = value.downcast::<String>() {
        println!("String ({}): {}", string.len(), string);
    }
}

let my_string = "Hello World".to_string();
let my_string: Rc<dyn Any + Send + Sync> = Rc::new(my_string); 
print_if_string(my_string.try_into().unwrap());
let my_number: Rc<dyn Any + Send + Sync> = Rc::new(0i8);
print_if_string(my_number.try_into().unwrap());

The unsizing as Rc is required until DST coercions are stabilized. Alternatively, activate the unsize feature to convert the pointer via an explicit method call.

impl<T: ?Sized> RcBox<T>

pub unsafe fn from_raw(ptr: *const T) -> Self

Construct an RcBox from a raw pointer.

Safety

The raw pointer must have previously been acquired by a call to RcBox::into_raw, or Rc::into_raw where the Rc is known unique.

pub fn get_mut(this: &mut Self) -> Option<&mut T>

๐Deprecated: Use DerefMut instead

Get a mutable reference into the RcBox.

This method exists only for API compatibility with Rc. Use DerefMut instead.

pub fn get_mut_unchecked(this: &mut Self) -> &mut T

๐Deprecated: Use DerefMut instead

Get a mutable reference into the RcBox.

This method exists only for API compatibility with Rc. Use DerefMut instead.

pub fn as_raw(this: &Self) -> NonNull<T>

Returns a raw pointer to the object T pointed to by this RcBox.

Note that this returns a ptr::NonNull, not a raw pointer. That makes this function equivalent to as_raw_non_null.

pub fn into_raw(this: Self) -> NonNull<T>

Consume the RcBox, returning the wrapped pointer.

To avoid a memory leak, the pointer must be converted back to a RcBox, using RcBox::from_raw, or directly into a Rc, using Rc::from_raw.

Note that this returns a ptr::NonNull, not a raw pointer. That makes this function equivalent to into_raw_non_null.

pub fn leak<'a>(this: Self) -> &'a mut T
where T: 'a,

Consume and leak the RcBox.

pub fn new(data: T) -> Self
where T: Sized,

Create a new RcBox.

pub fn pin(x: T) -> Pin<RcBox<T>>
where T: Sized,

Construct a new Pin<RcBox<T>>. If T does not implement Unpin, then the data will be pinned in memory and unable to be moved.

pub fn into_inner(this: Self) -> T
where T: Sized,

Deconstruct this RcBox, returning the inner value.